cli: Refactor to reduce cyclometric complexity
changelog: none

+++ cli/snapshot.go
@@ -32,6 +32,7 @@ import (
 )
 
 const (
+	// Error messages //
 	errMsgDataPartFmt = "Device-local data is stored on the rootfs " +
 		"partition: %s. The recommended approach is to have  a " +
 		"separate data-partition mounted on \"/data\" and add a " +
@@ -41,17 +42,16 @@ const (
 		"running `fsfreeze -u %s` or press `SYSRQ+j`, immediately!"
 
 	// Watchdog constants //
-
-	// WDTExpired is set by the listening routine if the timer expires.
-	WDTExpired int32 = -1
-	// WDTReset is swapped in on read by the listening routine.
-	WDTReset int32 = 0
-	// WDTSet is set by the monitored process.
-	WDTSet int32 = 1
-	// WDTIntervalSec Sets the polling interval of the listening routine.
+	// wdtExpired is set by the listening routine if the timer expires.
+	wdtExpired int32 = -1
+	// wdtReset is swapped in on read by the listening routine.
+	wdtReset int32 = 0
+	// wdtSet is set by the monitored process.
+	wdtSet int32 = 1
+	// wdtIntervalSec Sets the polling interval of the listening routine.
 	// NOTE: this should be sufficient enough for the user to for example
 	//       type their password to ssh etc.
-	WDTIntervalSec = 30
+	wdtIntervalSec = 30
 )
 
 var (
@@ -87,86 +87,89 @@ func (runOpts *runOptionsType) DumpSnapshot(ctx *cli.Context) error {
 	return runOpts.CopySnapshot(ctx, os.Stdout)
 }
 
-// CopySnapshot freezes the filesystem and copies a snapshot to out.
-func (runOpts *runOptionsType) CopySnapshot(ctx *cli.Context, out io.Writer) error {
+func checkSnapshotPreconditions(
+	fsPath, dataPath string) (*system.MountInfo, error) {
 	var err error
-	var fsSize uint64
 
-	// Ensure we don't write logs to the filesystem
-	if ctx.Bool("quiet") {
-		log.SetLevel(log.ErrorLevel)
-	}
-	log.SetOutput(os.Stderr)
-
-	if ctx.GlobalString("compression") == "gzip" {
-		out = gzip.NewWriter(out)
-	}
-
-	dataDevID, err := system.GetDeviceIDFromPath(runOpts.dataStore)
+	dataDevID, err := system.GetDeviceIDFromPath(dataPath)
 	if err != nil {
-		return err
+		return nil, err
 	}
-	rootDevID, err := system.GetDeviceIDFromPath(ctx.String("fs-path"))
+	rootDevID, err := system.GetDeviceIDFromPath(fsPath)
 	if err != nil {
-		return err
+		return nil, err
 	}
 	if dataDevID == rootDevID {
 		log.Errorf(errMsgDataPartFmt,
-			runOpts.dataStore, runOpts.dataStore)
-		return errors.Errorf(
-			"Data store (%s) is located on rootfs partition",
-			runOpts.dataStore)
+			dataPath, dataPath)
+		return nil, errors.Errorf(
+			"data store (%s) is located on rootfs partition",
+			dataPath)
 	}
 
-	var watchDog int32
 	rootDev, err := system.GetMountInfoFromDeviceID(rootDevID)
+	if err != nil {
+		return nil, err
+	} else if rootDev.FSType == "overlay" {
+		log.Error("overlay filesystem detected, " +
+			"filesystem type not supported")
+		log.Error("please specify the path to the " +
+			"block-device instead")
+		return nil, fmt.Errorf(
+			"snapshot: filesystem 'overlay' not supported")
+	}
+
+	return rootDev, nil
+}
+
+// CopySnapshot freezes the filesystem and copies a snapshot to out.
+func (runOpts *runOptionsType) CopySnapshot(ctx *cli.Context, out io.Writer) error {
+
+	var fd *os.File
+	var err error
+	var watchDog int32
+
+	// Ensure we don't write logs to the filesystem
+	log.SetOutput(os.Stderr)
+	if ctx.Bool("quiet") {
+		log.SetLevel(log.ErrorLevel)
+	}
+
+	rootDev, err := checkSnapshotPreconditions(
+		ctx.String("fs-path"),
+		ctx.GlobalString("data"),
+	)
 	if err == system.ErrDevNotMounted {
 		// If not mounted, assume the path points to a device
 	} else if err != nil {
 		return err
 	} else {
+		var f *os.File
 		sigChan := make(chan os.Signal)
 		abortChan := make(chan struct{})
-		signal.Notify(sigChan)
-		// freezeHandler is a transparent signal handler that ensures
-		// system.ThawFs is called upon a terminating signal.
-		// fsPath must be a directory in order to use the FIFREEZE ioctl,
-		// moreover, we don't have to freeze an unmounted device.
-		go freezeHandler(sigChan, abortChan,
-			rootDev.MountPoint, &watchDog)
-		defer func() {
-			// Check if handler has already returned
-			var hasAborted bool = false
-			select {
-			case _, hasAborted = <-abortChan:
-
-			default:
-			}
-			if !hasAborted {
-				// Terminate signal handler.
-				sigChan <- unix.SIGUSR1
-				// Ensure that the signal handler returns first.
-				<-abortChan
-			}
-			close(sigChan)
-			close(abortChan)
-		}()
-		if f, err := os.OpenFile(rootDev.MountPoint, 0, 0); err != nil {
+		if f, err = os.OpenFile(rootDev.MountPoint, 0, 0); err != nil {
+			defer f.Close()
+			// freezeHandler is a transparent signal handler that
+			// ensures system.ThawFs is called upon a terminating
+			// signal.
+			signal.Notify(sigChan)
+			go freezeHandler(sigChan, abortChan, f, &watchDog)
+			defer stopFreezeHandler(sigChan, abortChan)
 			if err == nil {
 				err = system.FreezeFS(int(f.Fd()))
 			}
-			if err != nil {
-				log.Warnf("Failed to freeze filesystem on %s: %s",
-					rootDev.MountSource, err.Error())
-				log.Warn("The snapshot might become invalid.")
-				abortChan <- struct{}{} // abort handler
-				signal.Stop(sigChan)
-			}
-
+		}
+		if err != nil {
+			log.Warnf("Failed to freeze filesystem on %s: %s",
+				rootDev.MountSource, err.Error())
+			log.Warn("The snapshot might become invalid.")
+			abortChan <- struct{}{} // abort handler
+			stopFreezeHandler(sigChan, abortChan)
+			signal.Stop(sigChan)
 		}
 	}
 
-	var fd *os.File
+	// Get filesystem size
 	if rootDev == nil {
 		fd, err = os.Open(ctx.String("fs-path"))
 	} else {
@@ -179,8 +182,12 @@ func (runOpts *runOptionsType) CopySnapshot(ctx *cli.Context, out io.Writer) err
 	}
 	defer fd.Close()
 
+	if ctx.GlobalString("compression") == "gzip" {
+		out = gzip.NewWriter(out)
+	}
+
 	// Get file system size - need to do this the hard way (returns uint64)
-	fsSize, err = system.GetBlockDeviceSize(fd)
+	fsSize, err := system.GetBlockDeviceSize(fd)
 	if err != nil {
 		return errors.Wrap(err, "Unable to get partition size")
 	}
@@ -214,24 +221,32 @@ func (runOpts *runOptionsType) CopySnapshot(ctx *cli.Context, out io.Writer) err
 // handler. The abort chan, as the name implies, aborts the handler without
 // executing FITHAW. This channel is also used to notify that the handler has
 // returned.
-func freezeHandler(sigChan chan os.Signal, abortChan chan struct{}, fsPath string, wdt *int32) {
+func freezeHandler(
+	sigChan chan os.Signal,
+	abortChan chan struct{},
+	fd *os.File,
+	wdt *int32,
+) {
 	var sig os.Signal = nil
+	var sigOpen bool = true
+	var abortOpen bool = true
 	for {
 		select {
-		case <-time.After(WDTIntervalSec * time.Second):
-			if old := atomic.SwapInt32(wdt, WDTReset); old > 0 {
+		case <-time.After(wdtIntervalSec * time.Second):
+			if old := atomic.SwapInt32(wdt, wdtReset); old > 0 {
 				continue
 			}
 			// Timer expired
-			atomic.StoreInt32(wdt, WDTExpired)
+			atomic.StoreInt32(wdt, wdtExpired)
 			log.Error("Watchdog timer expired due to " +
 				"blocked main process.")
 			log.Info("Unfreezing filesystem")
 
-		case <-abortChan:
+		case _, abortOpen = <-abortChan:
 			break
 
-		case sig = <-sigChan:
+		case sig, sigOpen = <-sigChan:
+
 		}
 		if sig != nil {
 			log.Debugf("Freeze handler received signal: %s",
@@ -247,26 +262,72 @@ func freezeHandler(sigChan chan os.Signal, abortChan chan struct{}, fsPath strin
 		}
 		// Terminating condition met (signal or closed channel
 		// -> Unfreeze rootfs
-		if f, err := os.OpenFile(fsPath, 0, 0); err != nil {
-			if err == nil {
-				err = system.FreezeFS(int(f.Fd()))
-			}
-			if err != nil {
-				log.Errorf(errMsgThawFmt, fsPath)
-			}
+		log.Debugf("Thawing filesystem at: %s", fd.Name())
+		if err := system.ThawFS(int(fd.Fd())); err != nil {
+			log.Errorf(errMsgThawFmt, fd.Name())
 		}
-		signal.Stop(sigChan)
-		if sig != nil && sig != unix.SIGUSR1 {
-			// Invoke default signal handler
-			unix.Kill(os.Getpid(), sig.(unix.Signal))
+		if sigOpen {
+			signal.Stop(sigChan)
+			if sig != nil && sig != unix.SIGUSR1 {
+				// Invoke default signal handler
+				unix.Kill(os.Getpid(), sig.(unix.Signal))
+			}
 		}
 		break
 	}
-	// Notify that the routine quit
-	abortChan <- struct{}{}
+	if abortOpen {
+		// Notify that the routine quit
+		abortChan <- struct{}{}
+	}
 }
 
-func CopyWithWatchdog(dst io.Writer, src io.Reader, wdt *int32, pb *utils.ProgressBar) error {
+// stopFreezeHandler ensures freezeHandler stops with the appropriate
+// preconditions
+func stopFreezeHandler(sigChan chan os.Signal, abortChan chan struct{}) {
+	// Check if handler has already returned
+	var abortOpen bool = true
+	var sigOpen bool = true
+	select {
+	case _, sigOpen = <-sigChan:
+
+	default:
+		signal.Stop(sigChan)
+	}
+	select {
+	case _, abortOpen = <-abortChan:
+		if sigOpen {
+			close(sigChan)
+		}
+		if abortOpen {
+			// The routine has already signaled abort
+			close(abortChan)
+		}
+
+	default:
+		if sigOpen {
+			// Both channels are open
+			sigChan <- unix.SIGUSR1
+			select {
+			// Wait no longer than a second (should not take long)
+			case <-time.After(time.Second):
+
+			case <-abortChan:
+			}
+			close(sigChan)
+		}
+		// Close abortChan regardless
+		close(abortChan)
+	}
+}
+
+// CopyWithWatchdog is an implementation of io.Copy that for each block it
+// copies resets the watchdog timer.
+func CopyWithWatchdog(
+	dst io.Writer,
+	src io.Reader,
+	wdt *int32,
+	pb *utils.ProgressBar,
+) error {
 	// Only tick every 10 writes
 	const tickInterval uint64 = 32 * 1024 * 10
 	var numTicks uint64
@@ -294,8 +355,8 @@ func CopyWithWatchdog(dst io.Writer, src io.Reader, wdt *int32, pb *utils.Progre
 				return err
 			}
 		}
-		wd := atomic.SwapInt32(wdt, WDTSet)
-		if wd == WDTExpired {
+		wd := atomic.SwapInt32(wdt, wdtSet)
+		if wd == wdtExpired {
 			return errWatchDogExpired
 		}
 		numTicks += uint64(n)
